+++ src/shared/hooks/search/useCampingSearch.js
